requirements.txt
fastapi
uvicorn
pytesseract
Pillow
python-multipart
boto3
pymysql
python-multipart
